requirements.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
certifi==2017.4.17
chardet==3.0.3
Django==1.11.1
django-celery-results==1.0.1
djangorestframework==3.6.3
idna==2.5
kombu==4.0.2
packaging==16.8
pymongo==3.4.0
pyparsing==2.2.0
pytz==2017.2
requests==2.16.5
six==1.10.0
urllib3==1.21.1
vine==1.1.3